Renamed ErrorSchemaValidation to SchemaValidation, added a NET35 constant to the VS2008 project and employed use for it in the NoticeComponentsCreation.Notice_contains_Request test. Also added test for Notice.Request.Action.
