+++ Tests/NoticeComponentsCreation.cs
@@ -54,10 +54,7 @@ namespace SharpBrake.Tests
         }
 
 
-        [Test(
-            Description =
-                "This test will fail in Visual Studio 2008 since we can't reference HttpSimulator from .NET 3.5."
-            )]
+        [Test]
         public void Notice_contains_Request()
         {
             AirbrakeNotice notice = null;
@@ -89,9 +86,14 @@ namespace SharpBrake.Tests
 
             Assert.That(notice, Is.Not.Null);
             Assert.That(notice.Error, Is.Not.Null);
+
+#if !NET35
+            // We have defined a NET35 constant in the Visual Studio 2008 project so the below code isn't executed,
+            // since it requires HttpSimulator which in turn requires .NET 4.0, which in turn requires Visual Studio 2010.
             Assert.That(notice.Request, Is.Not.Null);
             Assert.That(notice.Request.Url, Is.EqualTo(url));
             Assert.That(notice.Request.Component, Is.EqualTo(GetType().FullName));
+            Assert.That(notice.Request.Action, Is.EqualTo("Notice_contains_Request"));
 
             Assert.That(notice.Request.CgiData,
                         Contains.Item(new AirbrakeVar("Content-Type", "application/x-www-form-urlencoded")));
@@ -109,6 +111,7 @@ namespace SharpBrake.Tests
             Assert.That(notice.Request.Params, Contains.Item(new AirbrakeVar("Form.Key2", "Form.Value2")));
             Assert.That(notice.Request.Params, Contains.Item(new AirbrakeVar("Query.Key1", "Query.Value1")));
             Assert.That(notice.Request.Params, Contains.Item(new AirbrakeVar("Query.Key2", "Query.Value2")));
+#endif
         }
