+++ Tests/Tests.2008.csproj
@@ -1,6 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="3.5" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
+    <DefineConstants>NET35</DefineConstants>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProductVersion>9.0.30729</ProductVersion>
@@ -18,7 +19,7 @@
     <DebugType>full</DebugType>
     <Optimize>false</Optimize>
     <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
+    <DefineConstants>DEBUG;TRACE;NET35</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
@@ -26,7 +27,7 @@
     <DebugType>pdbonly</DebugType>
     <Optimize>true</Optimize>
     <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
+    <DefineConstants>TRACE;NET35</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
@@ -56,10 +57,10 @@
   <ItemGroup>
     <Compile Include="AirbrakeClientTests.cs" />
     <Compile Include="AirbrakeValidator.cs" />
-    <Compile Include="ErrorSchemaValidation.cs" />
     <Compile Include="HttpSimulatorStub.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="NoticeComponentsCreation.cs" />
+    <Compile Include="SchemaValidation.cs" />
     <Compile Include="Thrower.cs" />
     <Compile Include="XmlGeneration.cs" />
   </ItemGroup>
